Do not fetch user status if current user is a guest
The user status endpoint is not available for guests, only for
logged in users. As the request will always fail if the current user is
a guest there should be no need to perform it.

Note, however, that this will prevent user statuses to be shown on the
avatars in public share pages, even if the current user is actually
logged in. Public share pages are opened in ncognito mode, so
"getCurrentUser" returns null. However, if the current user is logged in
requests to the API would work without problems, so technically it would
be possible to fetch the user statuses.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>