Allow an "extra" slot for the AppNavigationItem
If you want to mount a Modal for an item in the navigation it becomes
tricky to mount the (virtual or not in-place) modal component from the
navigation. To avoid more complexity itwold be convenient to just mount
the component inside the navigation with the intention that it shouldn't
really render anything in-place.

Signed-off-by: Christoph Wurst <christoph@winzerhof-wurst.at>