Delay Popover display to avoid positioning glitches
Whenever the popover is set to "open" first, if the contents hasn't
already been populated in the current tick, the popover's position will
appear in the wrong location. Then, directly in the next tick its
position is corrected based on the newly appeared content.

To avoid the visual glitch we first keep it with "visibility: hidden"
upon opening, and then display it once its position has been calculated.

Signed-off-by: Vincent Petry <vincent@nextcloud.com>