Add `buttonTitle` property to ActionButton and PopoverMenuItem
The HTML button element supports the `title` property for tooltips.
Let's allow to set it in the ActionButton and PopoverMenuItem
components.

The name is `buttonTitle` instead of `title` because ActionButton
already has a `title` property with a different meaning and in
PopoverMenuItem elements other than button are supported. So let's be
explicit that this property is *only* for the `title` property of
`button` elements.

Signed-off-by: Jonas Meurer <jonas@freesources.org>