Fix box shadow colour of "tertiary-on-primary" button
The "tertiary-on-primary" button has a transparent background and
border, and it is meant to be shown on an element with primary colour.
Due to this the colour of the box shadow should be the primary text
colour rather than the default main text colour. Otherwise it would
change depending on light / dark mode, even if the button text and the
background it is shown on do not.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>