Do not treat an avatar as a button if there is no menu
An avatar is treated as a button that toggles its contact menu, except
when the menu was explicitly disabled. However, a menu may not be
available even if it was not explicitly disabled (for example, if the
avatar is for the current user). As no menu can be toggled in that case
the avatar should not be treated as a button either.

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>