Fix selectTemplate when there is no match
When no match was found for the current mention (that is, after pressing
"Enter" once the autocompletion list was hidden) "getnSelectTemplate"
receives an undefined value, which caused an exception when trying to
get "this.userData[value]". Now the same text that did not match is
returned in that case (this is based on the default "selectTemplate"
from Tribute).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>