fix: Render actions When using the composition API
The setup function of the composition API does not pass components by name but by instance to the vue render function `h`.
This results in `componentOptions.tag` being undefined, so for this case we need to check the `Ctor` instead.

Added an unit test to ensure this issue is fixed and we do not reintroduce this later as a regression.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>