Split Bootstrap version reference in Tooltip .scss
Bootstrap 3.3.5 has XSS vulnerabilities and some security scanners
report that Nextcloud is using a vulnerable dependency. However, this is
a false positive, as only some SCSS from that version is included and
nothing can be exploited.

It seems that the scanners just look for the version string instead of
actually checking the vulnerability, so the version reference was split
to avoid that (although the version is not simply removed to keep the
proper copyright assignment).

Signed-off-by: Daniel Calviño Sánchez <danxuliu@gmail.com>