feat(NcInputField): Make inner label always visible
The inner label is now visible, to disable any label set `label-outside` to true, `label-visible` property is removed.
The label is shown instead of the placeholder if the input is empty and not focussed, when focussed the label is moved up and scaled but always shown.
Make label color max-contrast so it does not look like content and remove horizontal label movement

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>