fix!(NcActionInput): Fix option to have the label next to the input field
In previous versions the label was shown above the input which gives the
text more space which might be better than the inline label from
https://github.com/nextcloud-libraries/nextcloud-vue/pull/4394 in some
scenarios.

This is a breaking change as it renames the previous labelVisible
property to a labelOutside property. The label should always be visible
with the new NcInputField, however we can optionally decide to have the
label outside.

Signed-off-by: Julius Härtl <jus@bitgrid.net>