fix(NcDialog): Adjust styles to prevent two scrollbars on small screens
This also fixes some design issues, we do not want padding or margin on the bottom and right of the content in case of a scrollbar.
When scrolling the scrollbar should be on the very left (not floating in the middle) and the scrolled content should be clipped by the modal container.

Also ensure the dialog name is not overflown by the close button.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>