fix(NcHeaderMenu): Remove padding from popover menu
There should be no padding within the popover menu.
The popover container is not styleable (at least not without hacking the (not exposed-)CSS classes),
so users can not remove the padding themself.

The padding is a problem if the content is scrollable,
because the scrollbar should be on the very end of the container and not floating around.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>