feat: Add `spawnDialog` function
This allows to spawn a Vue dialog without the need of mounting the
component.
So it is possible to spawn a dialog directly from any JS code outside of
components (e.g. callback functions).

This method was originally taken from `@nextcloud/dialogs`,
but I think it should belong to this Vue components library.

Co-authored-by: Ferdinand Thiessen <opensource@fthiessen.de>
Co-authored-by: Grigorii K. Shartsev <me@shgk.me>
Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>