feat(NcColorPicker): add bi-directional `open` prop
There was undocumented usage of the underlying `NcPopover` `shown` prop.
Instead this now provides the offical `open` prop for this usage.

The name was choosen because `open` is more consistently used in our
codebase for this use case instead of `shown`.

This also reverts the breaking changes entry for the `update:open` event.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>