refactor(NcButton): use proper prop type for `pressed` to allow `undefined`
For Vue a prop defined as `prop?: boolean` means that `prop` is a
boolean property with default `false`. But in this case we have a
tri-state value: either pressed (a boolean) or undefined.
So to allow such a prop to have the default value `undefined` instead of
`false` we need to explicitly type it as `prop?: boolean | undefined`.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>