fix(NcAppContentDetailsToggle): Hide navigation toggle on mobile
On mobile, the navigation toggle (`NcAppNavigationToggle`) needs to get
hidden when the app details toggle (`NcAppContentDetailsToggle`) is shown.

The code to do this was broken, as the watcher got triggered before the
navigation toggle was fully mounted, so `document.querySelector()`
didn't find the corresponding element.

Required to fix https://github.com/nextcloud/collectives/issues/2059

Signed-off-by: Jonas <jonas@freesources.org>