fix(NcPopover): Run `checkTriggerA11y()` only when dropdown is shown
`tabbable(<element>)` doesn't return elements that are hidden. This
makes the a11y test warn if the trigger element is hidden while the
component gets mounted.

E.g. in Nextcloud Text the menubar gets initialized with `visibility:
hidden`.

This changes the behaviour to only run `checkTrickerA11y()` when the
dropdown is shown.

Signed-off-by: Jonas <jonas@freesources.org>