build: fix translations in vue modules
When a vue file is handled by vite it will have some query string like
`source.vue?lang=ts&...` so to properly add the correct translations for
that file we need to remove the query string.

This resolves an issue where some files (e.g. `NcAppSettingsShortcutsSection`)
were missing their translations in the bundle.

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>